Minor refactoring
19 August update
